requirements.txt
fastapi==0.46.0
pydantic==1.10.2
uvicorn==0.11.1
redis==4.3.4
SQLAlchemy==1.4.19
python-dotenv==0.18.0
alembic==1.6.5
databases==0.6.1
asyncpg==0.26.0
psycopg2-binary==2.9.4
email-validator==1.3.0
fastapi-pagination==0.10.0